requirements.txt
Flask==2.0.2
openai==0.27.0
mysql-connector-python==8.0.26
Werkzeug==2.0.2